backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-ollama>=0.0.1
langchain-mongodb>=0.1.0
motor>=3.3.0
beautifulsoup4>=4.12.0
pydantic[email]>=2.0.0
python-multipart>=0.0.9
pypdf>=4.0.0
requests>=2.31.0
GitPython>=3.1.0
certifi>=2024.02.02
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
